fix(aec_bim, aec_bridge): external-IFC arg-2 fallback + BimFileSizeCheckJs f64 (Devin Review BUG_0001 + BUG_0002)
Devin Review findings on PR-O (#38):

BUG_0001 — napi_api.rs BimFileSizeCheckJs used BigInt for byte
counts, but the TypeScript interface in bridge.ts / preload.ts
declares 'number'. When the method moves from NATIVE_FALLBACK to
NATIVE_WIRED, the JS side would receive BigInt where it expects
Number — every arithmetic comparison ('fileSizeBytes >=
thresholdBytes') and division ('/ (1024 * 1024)') would throw
TypeError. Fix: change the struct to 'f64' (JS Number). f64 has
exact integer precision up to 2^53 (≈ 9 PB) which is many orders
of magnitude beyond any conceivable IFC file. This now matches
the BimImportSummaryJs convention (u32 / number for entity
counts) and the TS interface declarations.

BUG_0002 + ANALYSIS_0001 — reader.rs probed only arg 3
(IfcRoot.Description) for both spatial-node names and the
{tag}::{eid} element encoding. When real-world Revit / ArchiCAD
/ Tekla / buildingSMART exporters emit '$' (not-provided) at
arg 3 — which is the dominant convention since they put the
display string at arg 2 (IfcRoot.Name) — the reader propagated
an IfcReadError::Malformed up to the caller, silently dropping
the *entire* file. Now:

  * For spatial nodes (IFCPROJECT / IFCSITE / IFCBUILDING /
    IFCBUILDINGSTOREY / IFCSPACE), the name extraction is a
    three-stage fallback: try arg 3 first (AEC Studio writer
    convention), then arg 2 (IFC standard convention), then
    the empty string (so a node with '$' at both slots is
    still captured with no display label rather than aborting
    the whole parse).

  * For elements (the 'other' STEP-tag arm), the
    AEC-Studio-authored path-(a) detection now relies on
    'string_arg(3).ok().and_then(rsplit_once)' returning None
    when arg 3 is '$' — which correctly falls through to
    path (b) (classify by STEP tag, derive EntityId from
    GlobalId). Previously the entire branch was gated on
    'if let Ok(name) = g.string_arg(3)' so when arg 3 was
    '$', both paths were skipped and the element was
    silently dropped.

Two new regression tests on hand-rolled minimal STEP fixtures
with '$' at arg 3 pin both halves of the fix end-to-end:

  * external_ifc_with_dollar_description_captures_elements_via_step_tag
  * external_ifc_with_dollar_description_uses_arg2_name_on_spatial_nodes

cargo test -p aec_bim → 160 tests pass (146 lib + 10 + 2 + 1 + 1)
cargo test -p aec_bridge → 54 tests pass (47 lib + 7 integration)
cargo clippy --workspace --all-targets -- -D warnings → clean
cargo fmt --all --check → clean
npm test (renderer) → 176 tests pass

+++ crates/aec_bridge/src/napi_api.rs
@@ -342,28 +342,36 @@ pub fn bim_import_ifc(path: String) -> Result<BimImportSummaryJs> {
 /// JS-facing cheap file-size check. Mirrors the renderer's
 /// `BimFileSizeCheck` interface in `apps/desktop/electron/bridge.ts`.
 ///
-/// `file_size_bytes` and `threshold_bytes` use the
-/// JS-safe-integer-friendly `BigInt` shape via napi's `BigInt`
-/// type to avoid the f64-precision-cliff at 2^53 bytes (≈ 9 PB);
-/// any realistic IFC file is many orders of magnitude below
-/// that, but the type discipline matches the renderer's expected
-/// numeric handling and removes one class of "u64 silently
-/// truncated to JS Number" footgun.
+/// `file_size_bytes` and `threshold_bytes` are exposed as `f64`
+/// (JS `number`) so the renderer can do arithmetic on them directly
+/// (`fileSizeBytes / (1024 * 1024)` for humanised MB strings,
+/// `fileSizeBytes >= thresholdBytes` for the large-file guard)
+/// without BigInt / Number incompatibility. `f64` has exact
+/// integer precision up to 2^53 ≈ 9 PB — well beyond any
+/// conceivable IFC file — and matches the TS interface declaration
+/// of `number` in `bridge.ts` and `preload.ts`.
+///
+/// Consistency: this deliberately follows the same pattern as
+/// [`BimImportSummaryJs`], which uses `u32` (safe `number`) for
+/// entity counts. When this method is wired to the native bridge
+/// (i.e. moved from `NATIVE_FALLBACK_METHODS` to
+/// `NATIVE_WIRED_METHODS`), the JS side receives `number`
+/// directly — no BigInt / Number coercion footgun.
 #[napi(object)]
 pub struct BimFileSizeCheckJs {
     pub path: String,
-    pub file_size_bytes: BigInt,
+    pub file_size_bytes: f64,
     pub large_file_warning: bool,
-    pub threshold_bytes: BigInt,
+    pub threshold_bytes: f64,
 }
 
 impl From<crate::service::BimFileSizeCheck> for BimFileSizeCheckJs {
     fn from(r: crate::service::BimFileSizeCheck) -> Self {
         Self {
             path: r.path,
-            file_size_bytes: BigInt::from(r.file_size_bytes),
+            file_size_bytes: r.file_size_bytes as f64,
             large_file_warning: r.large_file_warning,
-            threshold_bytes: BigInt::from(r.threshold_bytes),
+            threshold_bytes: r.threshold_bytes as f64,
         }
     }
 }
